Update source to compile against more recent version of MyGUI and Ogre. Temporary comment irrklang audio library use. Add a Constant class to get data dir in easier way as static function call in one place.

+++ source/PowerNet.h
@@ -95,7 +95,7 @@ public:
 	int getEfficiency(void);
 
 	//! Get number of independent subnets
-	size_t getSubnetCount(void);
+    unsigned int getSubnetCount(void);
 
 	//! Return all powerplants in a specific subnet
 	std::vector<boost::shared_ptr<Powerplant> > getSubnetPowerplants(int pSubnetIndex);
